requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlite3  # Built-in, but for clarity
sqlalchemy>=2.0.0
alembic>=1.13.0  # For future migrations
python-multipart>=0.0.9  # For file uploads

# Optional backends
pymysql>=1.1.0
psycopg2-binary>=2.9.9

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0

# Utilities
pyyaml>=6.0.1
logging>=0.4.9.1  # Built-in
asyncio  # Built-in

# Radio interfaces
pyserial>=3.5
